octeon: enable interrupt mechanism
This patch enables the interrupt mechanism to support VFIO MSI-X by
fixing the PCI bus master enable in the Octeon plugin. Additionally, it
introduces an MSI-X message handler to support all vectors. Note that
interrupts for MBOX are disabled, as it requires a parallel thread
mechanism for handling.

Type: feature

Signed-off-by: Satha Rao <skoteshwar@marvell.com>
Signed-off-by: Monendra Singh Kushwaha <kmonendra@marvell.com>
Change-Id: Ic5cecd4a30526a86c25e88cac84d4bc2a5c8fbd5
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/toolkits/vpp/+/139703
Reviewed-by: Nithin Kumar Dabilpuram <ndabilpuram@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>