octeon: fix crypto failures after session delete
This patch fix the crypto failure that occurs when adding a
Security Association (SA) with a different algorithm but the
same ID after an SA deletion.
The issue was caused by reusing the context cache, which
has now been fixed by flushing and invalidating the cache
during the SA deletion process.

Type: fix

Signed-off-by: Nithinsen Kaithakadan <nkaithakadan@marvell.com>
Change-Id: I2099c0f98cab3787f8581b4c81e305a9b517d8ab
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/toolkits/vpp/+/144731
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Nithinsen Kaithakadan